backend/requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==3.14.0
django-templated-mail==1.1.1
django-tinymce==3.6.1
django-tinymce4-lite==1.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
frozenlist==1.3.3
idna==3.4
jsmin==3.0.1
magic-filter==1.0.9
multidict==6.0.3
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
vboxapi==1.0
yarl==1.8.2
